Added ant movement and drawing removed Debug folder from Gitignore to allow access to exe file

+++ AntSim/Structs.h
@@ -4,6 +4,6 @@
 
 struct vec2D
 {
-	uint16_t x = 0;
-	uint16_t y = 0;
+	double x = 0;
+	double y = 0;
 };
